requirements.txt
crewai
crewai_tools
streamlit 
pillow
streamlit
torch
langchain_cohere
tensor